refactor: split rewrite.rs (3,778 lines) into 8-module directory
Extract the monolithic cmd/rewrite.rs into cmd/rewrite/{mod,types,rules,
engine,compound,handlers,hook,suggest}.rs. Each module owns its domain:
types (data structures), rules (64-rule table with drift-prevention test),
engine (try_rewrite + stripping helpers), compound (&&/||/;/| splitting),
handlers (cat/head/tail custom logic), hook (agent PreToolUse integration),
suggest (clap command + JSON output), mod (public API + run entry point).

Visibility tightened to pub(super) at intra-module boundaries; only
would_rewrite() and run() remain pub(crate). All 1,156 tests pass; no
clippy warnings; rustfmt clean.
